@@ -1,7 +1,8 @@
 {
   "name": "webrdp",
-  "description": "Web RDP",
-  "version": "0.1.1",
+  "description": "Web-based RDP client proxy",
+  "version": "1.0.0",
+  "type": "module",
   "homepage": "https://github.com/Mikej81/WebRDP",
   "author": {
     "name": "Mikej81",
@@ -14,55 +15,46 @@
   "bugs": {
     "url": "https://github.com/Mikej81/WebRDP/issues"
   },
-  "license": "GPL",
-  "licenses": [
-    {
-      "type": "GPL",
-      "url": "https://github.com/Mikej81/WebRDP/blob/master/LICENSE-MIT"
-    }
-  ],
+  "license": "GPL-3.0",
   "main": "app.js",
   "engines": {
-    "node": "0.12.7"
+    "node": ">=20.0.0"
   },
   "scripts": {
-    "start": "node index",
-    "test": "snyk test",
-    "standard": "standard --verbose",
-    "build": "grunt copy concat uglify",
-    "cleanmac": "find . -name '.DS_Store' -type f -delete"
+    "start": "node app.js",
+    "dev": "node --watch app.js",
+    "build": "node esbuild.config.js",
+    "lint": "eslint .",
+    "test": "vitest run",
+    "test:watch": "vitest"
   },
   "dependencies": {
-    "base64-img": "^1.0.3",
-    "basic-auth": "^2.0.0",
-    "canvas-prebuilt": "^1.4.0",
-    "colors": "^1.1.2",
-    "debug": "^2.6.9",
-    "express": "^4.15.5",
-    "express-session": "^1.15.6",
-    "express-socket.io-session": "^1.3.2",
-    "morgan": "^1.8.2",
-    "node-rdpjs": "^0.3.0",
-    "read-config": "^2.0.0",
-    "socket.io": "^1.7.4",
-    "validator": "^9.4.1"
+    "@electerm/rdpjs": "^1.0.0",
+    "basic-auth": "^2.0.1",
+    "debug": "^4.3.4",
+    "dotenv": "^16.4.5",
+    "express": "^4.21.1",
+    "express-rate-limit": "^7.4.1",
+    "express-session": "^1.18.1",
+    "helmet": "^8.0.0",
+    "jimp": "^0.22.12",
+    "morgan": "^1.10.0",
+    "socket.io": "^4.8.1",
+    "validator": "^13.12.0"
   },
   "devDependencies": {
-    "bithound": "^1.7.0",
-    "grunt": "^1.0.1",
-    "grunt-contrib-concat": "^1.0.1",
-    "grunt-contrib-copy": "^1.0.0",
-    "grunt-contrib-uglify": "^3.1.0",
-    "snazzy": "^7.0.0",
-    "snyk": "^1.41.1",
-    "standard": "^10.0.3"
+    "@eslint/js": "^9.14.0",
+    "esbuild": "^0.24.0",
+    "eslint": "^9.14.0",
+    "socket.io-client": "^4.8.1",
+    "supertest": "^7.0.0",
+    "vitest": "^2.1.4"
   },
-  "standard": {
-    "ignore": [
-      "client/*.js",
-      "client/js/*.js",
-      "src/*.js"
-    ]
-  },
-  "keywords": []
+  "keywords": [
+    "rdp",
+    "remote-desktop",
+    "websocket",
+    "express",
+    "html5"
+  ]
 }
