@@ -17,174 +17,135 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-(function () {
-	/**
-	 * Mouse button mapping
-	 * @param button {integer} client button number
-	 */
-  function mouseButtonMap (button) {
-    switch (button) {
-      case 0:
-        return 1
-      case 2:
-        return 2
-      default:
-        return 0
-    }
-  };
+import { $, elementOffset, locale } from './mstsc.js';
+import { scancode } from './keyboard.js';
+import { createCanvas } from './canvas.js';
 
-	/**
-	 * Mstsc client
-	 * Input client connection (mouse and keyboard)
-	 * bitmap processing
-	 * @param canvas {canvas} rendering element
-	 */
-  function Client (canvas) {
-    this.canvas = canvas
-		// create renderer
-    this.render = new Mstsc.Canvas.create(this.canvas)
-    this.socket = null
-    this.activeSession = false
-    this.install()
+function mouseButtonMap(button) {
+  switch (button) {
+    case 0:
+      return 1;
+    case 2:
+      return 2;
+    default:
+      return 0;
+  }
+}
+
+class Client {
+  constructor(canvas) {
+    this.canvas = canvas;
+    this.render = createCanvas(this.canvas);
+    this.socket = null;
+    this.activeSession = false;
+    this.install();
   }
 
-  Client.prototype = {
-    install: function () {
-      var self = this
-			// bind mouse move event
-      this.canvas.addEventListener('mousemove', function (e) {
-        if (!self.socket) return
+  install() {
+    const self = this;
 
-        var offset = Mstsc.elementOffset(self.canvas)
-        self.socket.emit('mouse', e.clientX - offset.left, e.clientY - offset.top, 0, false)
-        e.preventDefault || !self.activeSession()
-        return false
-      })
-      this.canvas.addEventListener('mousedown', function (e) {
-        if (!self.socket) return
+    this.canvas.addEventListener('mousemove', (e) => {
+      if (!self.socket) return;
+      const offset = elementOffset(self.canvas);
+      self.socket.emit('mouse', e.clientX - offset.left, e.clientY - offset.top, 0, false);
+      e.preventDefault();
+      if (!self.activeSession) return false;
+      return false;
+    });
 
-        var offset = Mstsc.elementOffset(self.canvas)
-        self.socket.emit('mouse', e.clientX - offset.left, e.clientY - offset.top, mouseButtonMap(e.button), true)
-        e.preventDefault()
-        return false
-      })
-      this.canvas.addEventListener('mouseup', function (e) {
-        if (!self.socket || !self.activeSession) return
+    this.canvas.addEventListener('mousedown', (e) => {
+      if (!self.socket) return;
+      const offset = elementOffset(self.canvas);
+      self.socket.emit('mouse', e.clientX - offset.left, e.clientY - offset.top, mouseButtonMap(e.button), true);
+      e.preventDefault();
+      return false;
+    });
 
-        var offset = Mstsc.elementOffset(self.canvas)
-        self.socket.emit('mouse', e.clientX - offset.left, e.clientY - offset.top, mouseButtonMap(e.button), false)
-        e.preventDefault()
-        return false
-      })
-      this.canvas.addEventListener('contextmenu', function (e) {
-        if (!self.socket || !self.activeSession) return
+    this.canvas.addEventListener('mouseup', (e) => {
+      if (!self.socket || !self.activeSession) return;
+      const offset = elementOffset(self.canvas);
+      self.socket.emit('mouse', e.clientX - offset.left, e.clientY - offset.top, mouseButtonMap(e.button), false);
+      e.preventDefault();
+      return false;
+    });
 
-        var offset = Mstsc.elementOffset(self.canvas)
-        self.socket.emit('mouse', e.clientX - offset.left, e.clientY - offset.top, mouseButtonMap(e.button), false)
-        e.preventDefault()
-        return false
-      })
-      this.canvas.addEventListener('DOMMouseScroll', function (e) {
-        if (!self.socket || !self.activeSession) return
+    this.canvas.addEventListener('contextmenu', (e) => {
+      if (!self.socket || !self.activeSession) return;
+      const offset = elementOffset(self.canvas);
+      self.socket.emit('mouse', e.clientX - offset.left, e.clientY - offset.top, mouseButtonMap(e.button), false);
+      e.preventDefault();
+      return false;
+    });
 
-        var isHorizontal = false
-        var delta = e.detail
-        var step = Math.round(Math.abs(delta) * 15 / 8)
+    // Standard wheel event (replaces deprecated DOMMouseScroll and mousewheel)
+    this.canvas.addEventListener('wheel', (e) => {
+      if (!self.socket || !self.activeSession) return;
+      const isHorizontal = Math.abs(e.deltaX) > Math.abs(e.deltaY);
+      const delta = isHorizontal ? e.deltaX : e.deltaY;
+      const step = Math.round(Math.abs(delta) * 15 / 8);
+      const offset = elementOffset(self.canvas);
+      self.socket.emit('wheel', e.clientX - offset.left, e.clientY - offset.top, step, delta > 0, isHorizontal);
+      e.preventDefault();
+      return false;
+    });
 
-        var offset = Mstsc.elementOffset(self.canvas)
-        self.socket.emit('wheel', e.clientX - offset.left, e.clientY - offset.top, step, delta > 0, isHorizontal)
-        e.preventDefault()
-        return false
-      })
-      this.canvas.addEventListener('mousewheel', function (e) {
-        if (!self.socket || !self.activeSession) return
+    window.addEventListener('keydown', (e) => {
+      if (!self.socket || !self.activeSession) return;
+      self.socket.emit('scancode', scancode(e), true);
+      e.preventDefault();
+      return false;
+    });
 
-        var isHorizontal = Math.abs(e.deltaX) > Math.abs(e.deltaY)
-        var delta = isHorizontal ? e.deltaX : e.deltaY
-        var step = Math.round(Math.abs(delta) * 15 / 8)
+    window.addEventListener('keyup', (e) => {
+      if (!self.socket || !self.activeSession) return;
+      self.socket.emit('scancode', scancode(e), false);
+      e.preventDefault();
+      return false;
+    });
 
-        var offset = Mstsc.elementOffset(self.canvas)
-        self.socket.emit('wheel', e.clientX - offset.left, e.clientY - offset.top, step, delta > 0, isHorizontal)
-        e.preventDefault()
-        return false
-      })
-
-			// bind keyboard event
-      window.addEventListener('keydown', function (e) {
-        if (!self.socket || !self.activeSession) return
-
-        self.socket.emit('scancode', Mstsc.scancode(e), true)
-
-        e.preventDefault()
-        return false
-      })
-      window.addEventListener('keyup', function (e) {
-        if (!self.socket || !self.activeSession) return
-
-        self.socket.emit('scancode', Mstsc.scancode(e), false)
-
-        e.preventDefault()
-        return false
-      })
-
-      return this
-    },
-		/**
-		 * connect
-		 * @param ip {string} ip target for rdp
-		 * @param domain {string} microsoft domain
-		 * @param username {string} session username
-		 * @param password {string} session password
-		 * @param next {function} asynchrone end callback
-		 */
-    connect: function (next) {
-			// compute socket.io path (cozy cloud integration)
-      var parts = document.location.pathname.split('/'),
-		       base = parts.slice(0, parts.length - 1).join('/') + '/',
-		       path = base + 'socket.io'
-
-			// start connection
-      var self = this
-			// this.socket = io(window.location.protocol + "//" + window.location.host, { "path": path}).on('rdp-connect', function() {
-      this.socket = io(window.location.protocol + '//' + window.location.host).on('rdp-connect', function () {
-				// this event can be occured twice (RDP protocol stack artefact)
-        console.log('[WebRDP] connected')
-        self.activeSession = true
-      }).on('rdp-bitmap', function (bitmap) {
-        console.log('[WebRDP] bitmap update bpp : ' + bitmap.bitsPerPixel)
-        self.render.update(bitmap)
-      }).on('headerBackground', function (data) {
-        document.getElementById('header').style.backgroundColor = data
-      }).on('header', function (data) {
-        document.getElementById('header').innerHTML = data
-      }).on('screencap', function() {
-        var canvasimg = document.getElementsByTagName('canvas')
-        var dataUrl = canvasimg[0].toDataURL()
-        self.socket.emit('savescreen', dataUrl)
-      }).on('rdp-close', function () {
-        next(null)
-        console.log('[WebRDP] close')
-        self.activeSession = false
-      }).on('rdp-error', function (err) {
-        next(err)
-        console.log('[WebRDP] error : ' + err.code + '(' + err.message + ')')
-        self.activeSession = false
-      })
-
-			// emit infos event
-      this.socket.emit('infos', {
-        screen: {
-          width: this.canvas.width,
-          height: this.canvas.height
-        },
-        locale: Mstsc.locale()
-      })
-    }
+    return this;
   }
 
-  Mstsc.client = {
-    create: function (canvas) {
-      return new Client(canvas)
-    }
+  connect(next) {
+    const self = this;
+    this.socket = io(window.location.protocol + '//' + window.location.host)
+      .on('rdp-connect', () => {
+        console.log('[WebRDP] connected');
+        self.activeSession = true;
+      })
+      .on('rdp-bitmap', (bitmap) => {
+        self.render.update(bitmap);
+      })
+      .on('headerBackground', (data) => {
+        document.getElementById('header').style.backgroundColor = data;
+      })
+      .on('header', (data) => {
+        document.getElementById('header').textContent = data;
+      })
+      .on('rdp-close', () => {
+        next(null);
+        console.log('[WebRDP] close');
+        self.activeSession = false;
+      })
+      .on('rdp-error', (err) => {
+        next(err);
+        console.log('[WebRDP] error : ' + err.code + '(' + err.message + ')');
+        self.activeSession = false;
+      });
+
+    this.socket.emit('infos', {
+      screen: {
+        width: this.canvas.width,
+        height: this.canvas.height
+      },
+      locale: locale()
+    });
   }
-})()
+}
+
+export function createClient(canvas) {
+  return new Client(canvas);
+}
+
+// Expose to global scope for HTML inline scripts
+window.WebRDP = { $, createClient };
